+++ src/Types/EncryptedPassportElement.php
@@ -6,7 +6,7 @@ namespace TelegramBot\Types;
  * Contains information about documents or other Telegram Passport elements shared with the bot by the user.
  * Class EncryptedPassportElement
  * @package TelegramBot\Types
- * @link https://core.telegram.org/bots/api#encryptedpassportelement
+ * @link    https://core.telegram.org/bots/api#encryptedpassportelement
  */
 class EncryptedPassportElement {
 	
@@ -24,10 +24,14 @@ class EncryptedPassportElement {
 	 */
 	public $data;
 	
-	/** @var string $phone_number Optional. User's verified phone number, available only for “phone_number” type */
+	/**
+	 * @var string $phone_number Optional. User's verified phone number, available only for “phone_number” type
+	 */
 	public $phone_number;
 	
-	/** @var string $email Optional. User's verified email address, available only for “email” type */
+	/**
+	 * @var string $email Optional. User's verified email address, available only for “email” type
+	 */
 	public $email;
 	
 	/**
@@ -57,4 +61,18 @@ class EncryptedPassportElement {
 	 *     be decrypted and verified using the accompanying EncryptedCredentials.
 	 */
 	public $selfie;
+	
+	/**
+	 * @var PassportFile[] $translation Optional. Array of encrypted files with translated versions of documents
+	 *      provided by the user. Available if requested for “passport”, “driver_license”, “identity_card”,
+	 *      “internal_passport”, “utility_bill”, “bank_statement”, “rental_agreement”, “passport_registration” and
+	 *      “temporary_registration” types. Files can be decrypted and verified using the accompanying
+	 *      EncryptedCredentials.
+	 */
+	public $translation;
+	
+	/**
+	 * @var string $hash Base64-encoded element hash for using in PassportElementErrorUnspecified
+	 */
+	public $hash;
 }